Removed percentage from the event. Moved event into the repo API.
