+++ hawkbit-repository/hawkbit-repository-api/src/main/java/org/eclipse/hawkbit/repository/ControllerManagement.java
@@ -14,8 +14,8 @@ import java.util.Map;
 
 import javax.validation.constraints.NotNull;
 
-import org.eclipse.hawkbit.eventbus.event.DownloadProgressEvent;
 import org.eclipse.hawkbit.im.authentication.SpPermission.SpringEvalExpressions;
+import org.eclipse.hawkbit.repository.eventbus.event.DownloadProgressEvent;
 import org.eclipse.hawkbit.repository.exception.EntityAlreadyExistsException;
 import org.eclipse.hawkbit.repository.exception.EntityNotFoundException;
 import org.eclipse.hawkbit.repository.exception.ToManyAttributeEntriesException;
@@ -58,21 +58,20 @@ public interface ControllerManagement {
     Action addCancelActionStatus(@NotNull ActionStatus actionStatus);
 
     /**
-     * Sends the download progress in percentage and notifies the
-     * {@link EventBus} with a {@link DownloadProgressEvent}.
+     * Sends the download progress and notifies the {@link EventBus} with a
+     * {@link DownloadProgressEvent}.
      * 
      * @param statusId
      *            the ID of the {@link ActionStatus}
-     * @param progressPercent
-     *            the progress in percentage which must be between 0-100
+     * @param requestedBytes
+     *            requested bytes of the request
      * @param shippedBytesSinceLast
      *            since the last report
      * @param shippedBytesOverall
      *            for the {@link ActionStatus}
      */
     @PreAuthorize(SpringEvalExpressions.IS_CONTROLLER)
-    void downloadProgressPercent(long statusId, int progressPercent, long shippedBytesSinceLast,
-            long shippedBytesOverall);
+    void downloadProgress(Long statusId, Long requestedBytes, Long shippedBytesSinceLast, Long shippedBytesOverall);
 
     /**
      * Simple addition of a new {@link ActionStatus} entry to the {@link Action}
