+++ hawkbit-repository/hawkbit-repository-jpa/src/main/java/org/eclipse/hawkbit/repository/jpa/model/JpaActionStatus.java
@@ -24,7 +24,10 @@ import javax.persistence.ManyToOne;
 import javax.persistence.NamedAttributeNode;
 import javax.persistence.NamedEntityGraph;
 import javax.persistence.Table;
+import javax.persistence.Transient;
 
+import org.eclipse.hawkbit.repository.jpa.cache.CacheField;
+import org.eclipse.hawkbit.repository.jpa.cache.CacheKeys;
 import org.eclipse.hawkbit.repository.model.Action;
 import org.eclipse.hawkbit.repository.model.Action.Status;
 import org.eclipse.hawkbit.repository.model.ActionStatus;
@@ -62,6 +65,13 @@ public class JpaActionStatus extends AbstractJpaTenantAwareBaseEntity implements
             @Index(name = "sp_idx_action_status_msgs_01", columnList = "action_status_id") })
     @Column(name = "detail_message", length = 512)
     private final List<String> messages = new ArrayList<>();
+
+    /**
+     * Note: filled only in {@link Status#DOWNLOAD}.
+     */
+    @Transient
+    @CacheField(key = CacheKeys.DOWNLOAD_PROGRESS_PERCENT)
+    private int downloadProgressPercent;
 
     /**
      * Creates a new {@link ActionStatus} object.
@@ -103,6 +113,11 @@ public class JpaActionStatus extends AbstractJpaTenantAwareBaseEntity implements
      */
     public JpaActionStatus() {
         // JPA default constructor.
+    }
+
+    @Override
+    public int getDownloadProgressPercent() {
+        return downloadProgressPercent;
     }
 
     @Override
