+++ hawkbit-repository/hawkbit-repository-jpa/src/test/java/org/eclipse/hawkbit/repository/jpa/cache/CacheWriteNotifyTest.java
@@ -13,7 +13,7 @@ import static org.mockito.Matchers.eq;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
-import org.eclipse.hawkbit.eventbus.event.DownloadProgressEvent;
+import org.eclipse.hawkbit.repository.eventbus.event.DownloadProgressEvent;
 import org.eclipse.hawkbit.repository.model.ActionStatus;
 import org.eclipse.hawkbit.tenancy.TenantAware;
 import org.junit.Before;
@@ -59,15 +59,14 @@ public class CacheWriteNotifyTest {
     @Test
     public void downloadgProgressIsCachedAndEventSent() {
         final long knownStatusId = 1;
-        final int knownPercentage = 23;
 
         when(cacheManagerMock.getCache(ActionStatus.class.getName())).thenReturn(cacheMock);
         when(tenantAwareMock.getCurrentTenant()).thenReturn("default");
 
-        underTest.downloadProgressPercent(knownStatusId, knownPercentage, 100L, 500L);
+        underTest.downloadProgress(knownStatusId, 500L, 100L, 100L);
 
         verify(cacheManagerMock).getCache(eq(ActionStatus.class.getName()));
-        verify(cacheMock).put(knownStatusId + "." + CacheKeys.DOWNLOAD_PROGRESS_PERCENT, knownPercentage);
+        verify(cacheMock).put(knownStatusId + "." + CacheKeys.DOWNLOAD_PROGRESS_PERCENT, 20);
         verify(eventBusMock).post(any(DownloadProgressEvent.class));
     }
